[refactoring] restructure modules to have responses alongside requests as top-level modules. move errors, records, and resources to responses.
