+++ src/app/application/projects/withEntityForm.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { createConnectionChecker } from '../shared/net/http/requests/connections/Repository';
+import { createConnectionChecker } from '../shared/net/http/requests/connections';
 import {
   createAfterUpdateResourcesLifeCycle,
   createPatchPayload,
