+++ src/app/application/shared/net/http/requests/connections/index.js
@@ -1,0 +1,1 @@
+export * from './Repository2';
