+++ src/app/application/shared/net/http/requests/index.js
@@ -2,4 +2,6 @@
  * @flow
  */
 
+export * from './Repository';
+export * from './Services';
 export * from './Types';
