+++ src/app/application/shared/net/http/responses/Types.js
@@ -1,0 +1,11 @@
+/*
+ * @flow
+ */
+
+import type { ErrorMap } from './errors/Types';
+import type { ResourceDatabase } from './resources/Types';
+
+export type ResponseState = {
+  +errors: ErrorMap,
+  +resources: ResourceDatabase
+};
