+++ src/app/application/shared/net/http/responses/index.js
@@ -1,0 +1,1 @@
+export * from './Repository';
