+++ src/app/application/shared/net/http/responses/records/Types.js
