+++ src/app/application/shared/net/http/responses/records/index.js
@@ -1,0 +1,1 @@
+export * from './Repository';
