+++ src/app/application/shared/net/http/responses/resources/Services.js
