+++ src/app/application/shared/net/http/responses/resources/Types.js
