+++ src/app/application/shared/net/http/responses/resources/index.js
@@ -1,2 +1,3 @@
 export * from './Repository';
 export * from './Services';
+export * from './Types';
