+++ src/app/application/shared/net/http/types.js
@@ -3,26 +3,6 @@
  */
 
 import type { RequestState } from './requests/Types';
-import type { ResourceDatabase } from './resources/Types';
+import type { ResponseState } from './responses/Types';
 
-//-----------------
-// BEGIN HTTP STATE
-//-----------------
-
-export type HttpState = { +requests: RequestState, +resources: ResourceDatabase };
-
-//---------------
-// END HTTP STATE
-//---------------
-
-//------------------
-// BEGIN HTTP ERRORS
-//------------------
-
-export type HttpError = { +detail: string };
-
-export type HttpErrorCollection = Array<HttpError> | null;
-
-//----------------
-// END HTTP ERRORS
-//----------------
+export type HttpState = { +requests: RequestState, +responses: ResponseState };
