+++ src/app/application/utils/withEntity.js
@@ -2,10 +2,10 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { createRecordGetter } from '../shared/net/http/records/Repository';
-import { hasQueryMetaResult } from '../shared/net/http/requests/queries/Repository';
-import { createConnectionChecker } from '../shared/net/http/requests/connections/Repository';
-import { createErrorGetter } from '../shared/net/http/requests/errors/Repository';
+import { createErrorGetter } from '../shared/net/http/responses/errors';
+import { createRecordGetter } from '../shared/net/http/responses/records';
+import { hasQueryMetaResult } from '../shared/net/http/requests/queries';
+import { createConnectionChecker } from '../shared/net/http/requests/connections';
 
 export default ({ readResource, requestId, resourceType }) => (
   (WrappedComponent) => {
